add buildEvent and buildError handlers, refactor handleCalendar
