+++ src/js/Handlers/handleCheckDate.js
@@ -1,13 +1,13 @@
 "use strict";
 
-import utilities from "../Utilities/Utilities";
+import Utilities from "../Utilities/Utilities";
 
 const handleCheckDate = monthEvent => {
-  var eventMonth = utilities.getMonthFromString(monthEvent[1]);
+  var eventMonth = Utilities.getMonthFromString(monthEvent[1]);
   var eventDay = Number(monthEvent[2].slice(0, -2));
 
-  var currentMonth = Number(utilities.getMonth());
-  var currentDay = Number(utilities.getDate());
+  var currentMonth = Number(Utilities.getMonth());
+  var currentDay = Number(Utilities.getDate());
 
   if (currentMonth < eventMonth) {
     return true;
